remove Render filed from common product
