+++ pkg/cli/upgradeassistant/internal/repository/mongodb/product.go
@@ -148,3 +148,13 @@ func (c *ProductColl) UpdateProductRender(product *models.Product) error {
 	_, err := c.Collection.UpdateOne(context.TODO(), query, change)
 	return err
 }
+
+func (c *ProductColl) UpdateRender(envName, productName string, render *models.RenderInfo) error {
+	query := bson.M{"env_name": envName, "product_name": productName}
+	change := bson.M{"$set": bson.M{
+		"render": render,
+	}}
+	_, err := c.UpdateOne(context.TODO(), query, change)
+
+	return err
+}
