+++ pkg/microservice/aslan/core/common/repository/mongodb/product.go
@@ -352,16 +352,6 @@ func (c *ProductColl) UpdateRegistry(envName, productName, registryId string) er
 	return err
 }
 
-func (c *ProductColl) UpdateRender(envName, productName string, render *models.RenderInfo) error {
-	query := bson.M{"env_name": envName, "product_name": productName}
-	change := bson.M{"$set": bson.M{
-		"render": render,
-	}}
-	_, err := c.UpdateOne(context.TODO(), query, change)
-
-	return err
-}
-
 func (c *ProductColl) Delete(owner, productName string) error {
 	query := bson.M{"env_name": owner, "product_name": productName}
 	_, err := c.DeleteOne(context.TODO(), query)
